+++ clients/derivatives-trading-coin-futures/src/rest-api/types/kline-candlestick-data-response-item.ts
@@ -14,14 +14,9 @@
  * Do not edit the class manually.
  */
 
-// May contain unused imports in some cases
-// @ts-ignore
-import type { ContinuousContractKlineCandlestickDataResponseItemInner } from './continuous-contract-kline-candlestick-data-response-item-inner';
-
 /**
  *
  * @export
  * @interface KlineCandlestickDataResponseItem
  */
-export interface KlineCandlestickDataResponseItem
-    extends Array<ContinuousContractKlineCandlestickDataResponseItemInner> {}
+export interface KlineCandlestickDataResponseItem extends Array<string> {}