+++ clients/derivatives-trading-coin-futures/tests/unit/rest-api/AccountApiTest.test.ts
@@ -702,10 +702,10 @@ describe('AccountApi', () => {
 
             mockResponse = {
                 downloadId: '545923594199212032',
-                status: 'completed',
-                url: 'www.binance.com',
-                notified: true,
-                expirationTimestamp: 1645009771000,
+                status: 'processing',
+                url: '',
+                notified: false,
+                expirationTimestamp: -1,
                 isExpired: null,
             };
 
@@ -733,10 +733,10 @@ describe('AccountApi', () => {
 
             mockResponse = {
                 downloadId: '545923594199212032',
-                status: 'completed',
-                url: 'www.binance.com',
-                notified: true,
-                expirationTimestamp: 1645009771000,
+                status: 'processing',
+                url: '',
+                notified: false,
+                expirationTimestamp: -1,
                 isExpired: null,
             };
 
@@ -800,10 +800,10 @@ describe('AccountApi', () => {
 
             mockResponse = {
                 downloadId: '545923594199212032',
-                status: 'completed',
-                url: 'www.binance.com',
-                notified: true,
-                expirationTimestamp: 1645009771000,
+                status: 'processing',
+                url: '',
+                notified: false,
+                expirationTimestamp: -1,
                 isExpired: null,
             };
 
@@ -829,10 +829,10 @@ describe('AccountApi', () => {
 
             mockResponse = {
                 downloadId: '545923594199212032',
-                status: 'completed',
-                url: 'www.binance.com',
-                notified: true,
-                expirationTimestamp: 1645009771000,
+                status: 'processing',
+                url: '',
+                notified: false,
+                expirationTimestamp: -1,
                 isExpired: null,
             };
 
@@ -894,10 +894,10 @@ describe('AccountApi', () => {
 
             mockResponse = {
                 downloadId: '545923594199212032',
-                status: 'completed',
-                url: 'www.binance.com',
-                notified: true,
-                expirationTimestamp: 1645009771000,
+                status: 'processing',
+                url: '',
+                notified: false,
+                expirationTimestamp: -1,
                 isExpired: null,
             };
 
@@ -925,10 +925,10 @@ describe('AccountApi', () => {
 
             mockResponse = {
                 downloadId: '545923594199212032',
-                status: 'completed',
-                url: 'www.binance.com',
-                notified: true,
-                expirationTimestamp: 1645009771000,
+                status: 'processing',
+                url: '',
+                notified: false,
+                expirationTimestamp: -1,
                 isExpired: null,
             };